scripts/requirements.txt
pandas
sqlalchemy
psycopg2-binary
pyarrow
s3fs
python-dotenv
kafka-python
scikit-learn
streamlit